requirements.txt
anyio==3.7.0
black==23.3.0
certifi==2023.5.7
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.98.0
filelock==3.12.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
identify==2.5.24
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.24.4
orjson==3.9.1
packaging==23.1
pandas==2.0.3
pathspec==0.11.1
platformdirs==3.8.0
pre-commit==3.3.3
psycopg2-binary==2.9.6
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
tomli==2.0.1
typing-extensions==4.7.0
tzdata==2023.3
ujson==5.8.0
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.23.1
watchfiles==0.19.0
websockets==11.0.3
